feat(integration): merge table creation logic
