+++ src/components/molecules/IntegrationMappingCard/index.ts
@@ -1,0 +1,1 @@
+export { default as IntegrationMappingCard } from './IntegrationMappingCard';
